pyproject.toml
[project]
name = "ml-final-cifar-100"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.179.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "ipykernel>=6.30.1",
    "numpy>=2.2.6",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "wandb>=0.21.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]

[tool.pytest.ini_options]
pythonpath = "src"
